requirements.txt
python-dotenv
fastapi
uvicorn
pydantic
sqlalchemy
aiomysql
asyncpg
pyjwt
passlib